Change bonus calculation type from double to decimal

+++ Services/EmployeeService.cs
@@ -3,14 +3,14 @@ namespace SistemaGestionEmpleados.Services;
 
 public class EmployeeService
 {
-    private double CalculateBonus(Employee employee)
+    private decimal CalculateBonus(Employee employee)
     {
         return employee.Salary * 0.10;
     }
 
     public void ShowInfo(Employee employee)
     {
-        double salaryPlusBonus = employee.Salary + CalculateBonus(employee);
+        decimal salaryPlusBonus = employee.Salary + CalculateBonus(employee);
         Console.WriteLine($"Employee's name: {employee.FirstName} {employee.LastName}");
         Console.WriteLine($"Employee's age: {employee.Age}");
         Console.WriteLine($"Employee's Role: {employee.Role}");
